Major refactoring
removed all warning and error code from vector3SIMD and added it in vector3.h. Vector3.h now completely controls the flow and does not rely on vector3SIMD.h somehow returning back to vector3.h. This is more efficient because vector3SIMD.h is never entered unless it's corresponding method that we are trying to access is well defined and functional on the machine we are using.

vector3SIMD.h breaks each method into an sse method and a neon method.