Completely revamped vector3 scripts
I had to change the naming structures of vector3.h so they would be recognized by the rest of the engine. That meant I had to change vector3.cpp and vector3SIMD as well. I also removed all the gatekeeping and error-checking logic from within vector3SIMD.h and entrusted that to vector3.h.